Move components out of lab
