+++ packages/web-ui/src/Link/Link.docs.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, ArgTypes } from '@storybook/blocks';
 import * as Stories from './Link.stories';
-import { DocsHeader } from '../../storybook-utils';
+import { DocsHeader } from '../storybook-utils';
 
 <Meta of={Stories} />
