+++ packages/web-ui/src/TextLink/TextLink.docs.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, ArgTypes } from '@storybook/blocks';
 import * as Stories from './TextLink.stories';
-import { DocsHeader } from '../../storybook-utils';
+import { DocsHeader } from '../storybook-utils';
 
 <Meta of={Stories} />
